requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
python-dotenv
pypdf
python-docx
requests
streamlit
